The Road Wardens Updates (#100)
- Replace using message with requires
RoadWardensMMM.esp is a patch for RoadWardens.esm and Mart's Mutant Mod.esm. 
It includes one change to a script. 
It requires RoadWardens.esp but does not have it as a master.

- Add afters for recommended order
Below is the load order guidelines, which can be found on the modpage description.
https://www.nexusmods.com/fallout3/mods/2781

1) RoadWardens.esm & Mart's Mutant Mod.esm (order not important)
2) Mart's Mutant Mod.esp
3) Optional Mart's Mutant Mods (like Increased Spawns and Hunting & Looting)
4) RoadWardens.esp
5) RoadWardensMMM.esp

I've tested in xEdit against all the plugins in the MMM RC 62 installer.
The only conflicts with RoadWardens.esp are:
Mart's Mutant Mod.esp and Mart's Mutant Mod - Tougher Traders.esp.

- Place mod plugins together
- Add Location
https://www.nexusmods.com/fallout3/mods/2781/